github: Move `From<GitHubError> for BoxedAppError` block to the `util/errors` module
The `github` module should not know anything about the HTTP layer.
